examples/mpu, loadable_sample/wifiapp: fix build warnings
The ioctl requests unsigned long type of argument.
Let's do type casting for that.
This commit fixes belows:

CC:  mpu_test.c
mpu_test.c: In function 'mpu_tc_main':
mpu_test.c:59:40: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
    ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                        ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
mpu_test.c:89:41: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
     ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                         ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
mpu_test.c:111:41: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
     ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                         ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
mpu_test.c:142:40: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
    ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                        ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
mpu_test.c:172:41: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
     ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                         ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
mpu_test.c:194:41: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
     ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                         ^
In file included from mpu_test.c:26:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~

CC: recovery.c
recovery.c: In function 'assert_thread':
recovery.c:55:39: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
   ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                       ^
In file included from recovery.c:30:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~
recovery.c:68:39: warning: passing argument 3 of 'ioctl' makes integer from pointer without a cast [-Wint-conversion]
   ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
                                       ^
In file included from recovery.c:30:0:
/root/tizenrt/os/include/sys/ioctl.h:140:5: note: expected 'long unsigned int' but argument is of type 'struct mputest_arg_s *'
 int ioctl(int fd, int req, unsigned long arg);
     ^~~~~

+++ apps/examples/mpu/mpu_test.c
@@ -56,7 +56,7 @@ int mpu_tc_main(int argc, char *argv[])
 
 		if (ch == 'A' || ch == 'a') {
 			obj.type = MPUTEST_APP_ADDR;
-			ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+			ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 			if (ret < 0) {
 				printf("ERROR: Failed to fetch test address from kernel\n");
 				close(tc_fd);
@@ -86,7 +86,7 @@ int mpu_tc_main(int argc, char *argv[])
 			ch = getchar();
 			if (ch == 'C' || ch == 'c') {
 				obj.type = MPUTEST_KERNEL_CODE;
-				ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+				ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 				if (ret < 0) {
 					printf("ERROR: Failed to fetch test address from kernel\n");
 					close(tc_fd);
@@ -108,7 +108,7 @@ int mpu_tc_main(int argc, char *argv[])
 				printf("ERROR: User Task made invalid access to Kernel code space\n");
 			} else if (ch == 'D' || ch == 'd') {
 				obj.type = MPUTEST_KERNEL_DATA;
-				ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+				ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 				if (ret < 0) {
 					printf("ERROR: Failed to fetch test address from kernel\n");
 					close(tc_fd);
@@ -139,7 +139,7 @@ int mpu_tc_main(int argc, char *argv[])
 
 		if (ch == 'A' || ch == 'a') {
 			obj.type = MPUTEST_APP_ADDR;
-			ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+			ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 			if (ret < 0) {
 				printf("ERROR: Failed to fetch test address from kernel\n");
 				close(tc_fd);
@@ -169,7 +169,7 @@ int mpu_tc_main(int argc, char *argv[])
 			ch = getchar();
 			if (ch == 'C' || ch == 'c') {
 				obj.type = MPUTEST_KERNEL_CODE;
-				ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+				ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 				if (ret < 0) {
 					printf("ERROR: Failed to fetch test address from kernel\n");
 					close(tc_fd);
@@ -191,7 +191,7 @@ int mpu_tc_main(int argc, char *argv[])
 				printf("ERROR: User Task made invalid access to Kernel code space\n");
 			} else if (ch == 'D' || ch == 'd') {
 				obj.type = MPUTEST_KERNEL_DATA;
-				ret = ioctl(tc_fd, TESTIOC_MPUTEST, &obj);
+				ret = ioctl(tc_fd, TESTIOC_MPUTEST, (unsigned long)&obj);
 				if (ret < 0) {
 					printf("ERROR: Failed to fetch test address from kernel\n");
 					close(tc_fd);
